# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Role {
  id: string;
  name: string;
  goal1: string;
  goal2: string;
}

export interface Schedule {
  السبت: string;
  الأحد: string;
  الاثنين: string;
  الثلاثاء: string;
  الأربعاء: string;
  الخميس: string;
  الجمعة: string;
}

export interface ReviewEntry {
  roleId: string;
  roleName: string;
  goalField: string;
  goalText: string;
  statusVal: string;
  reason: string;
  advice: string;
  concept: string;
}

export interface Review {
  entries: ReviewEntry[];
  message: string;
  completedAt: string;
}

export interface PlannerData {
  weekEndDate: string;
  roles: Role[];
  schedule: Schedule;
  reflection: string;
  review?: Review | null;
}

export interface PlannerDataInput {
  weekEndDate: string;
  roles: Role[];
  schedule: Schedule;
  reflection: string;
  review?: Review | null;
}

export interface ReviewAnalysisEntry {
  role: string;
  goal: string;
  status: string;
  /** @nullable */
  reason: string | null;
}

export interface ReviewAnalysisInput {
  entries: ReviewAnalysisEntry[];
}

export interface ReviewAnalysisItem {
  advice: string;
  concept: string;
}

export interface ReviewAnalysis {
  items: ReviewAnalysisItem[];
  message: string;
}