# lib/api-zod/src/generated/types/plannerDataInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Review } from './review';
import type { Role } from './role';
import type { Schedule } from './schedule';

export interface PlannerDataInput {
  weekEndDate: string;
  roles: Role[];
  schedule: Schedule;
  reflection: string;
  review?: Review | null;
}