# lib/api-zod/src/generated/types/reviewAnalysis.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewAnalysisItem } from './reviewAnalysisItem';

export interface ReviewAnalysis {
  items: ReviewAnalysisItem[];
  message: string;
}