# lib/api-zod/src/generated/types/reviewAnalysisEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewAnalysisEntry {
  role: string;
  goal: string;
  status: string;
  /** @nullable */
  reason: string | null;
}