# lib/api-zod/src/generated/types/reviewAnalysisInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewAnalysisEntry } from './reviewAnalysisEntry';

export interface ReviewAnalysisInput {
  entries: ReviewAnalysisEntry[];
}