# lib/api-zod/src/generated/types/schedule.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Schedule {
  السبت: string;
  الأحد: string;
  الاثنين: string;
  الثلاثاء: string;
  الأربعاء: string;
  الخميس: string;
  الجمعة: string;
}